Add os checks to bootstrap script

+++ bin/install-dota.sh
@@ -1,8 +1,28 @@
 #!/bin/sh
 
-apt-get update
-apt-get upgrade -y
-apt-get install -y puppet git-core
+if [ -f /etc/debian_version ]; then
+    echo "Detected Debian system, installing..."
+    apt-get update
+    apt-get upgrade -y
+    apt-get install -y puppet git-core
+elif [ -f /etc/redhat-release ]; then
+    if grep ' 5\.' /etc/redhat-release >/dev/null; then
+        RELEASE = 5
+    elif grep ' 6\.' /etc/redhat-release >/dev/null; then
+        RELEASE = 6
+    else
+        echo "Unsuported CentOS/RedHat version, please use 5.x or 6.x."
+        exit 1
+    fi
+    echo "Detected CentOS/RedHat system, installing..."
+    rpm -ivh http://yum.puppetlabs.com/el/5/products/i386/puppetlabs-release-$RELEASE-6.noarch.rpm
+    yum update
+    yum upgrade -y
+    yum install -y puppet git
+else
+    echo 'Unsuported operating system.  Email contact@barroncraft.com if you need help.'
+fi
+
 rm -rf /etc/puppet; mkdir /etc/puppet
 git clone git://github.com/barroncraft/barroncraft-puppet.git /etc/puppet
-puppet /etc/puppet/manifests/dotaSetup.pp
+puppet apply /etc/puppet/manifests/dotaSetup.pp
